modules/mavlinkModules/MAVLinkThread/requirements.txt
future==0.17.1
lxml==4.4.0
pymavlink==2.3.8
pyserial==3.4
